chore(core): reorganize string expression parser
This isn't actually about predicates, it's about evaluating expressions in order to pass them to the predicates. Cleaning this up makes the boundaries clearer.
